feat(kg): Custom(String) edge escape hatch + Option H warm-boot (closes #818, #816)
Implements ADR-0010 Option H / Phase E:

• `contracts::EdgeKind`: add `Custom(String)` variant (score=0.70), remove
  `Copy` (String is heap-allocated), widen `edge_kind_tag` to
  `Cow<'static, str>` via `tag()`/`from_tag()`/`from_static_tag()` methods.
  On-disk tag: `"custom:<s>"` — permanently reserved prefix per ADR-0010.
• Doc advisories: `Writes` (0.90) marked provisional/heuristic pending pilot
  data; `AccessesResource` precedence note added (prefer Reads/Writes when
  direction is known).
• `contracts.rs` → `contracts/` module directory: `mod.rs` (entity types,
  RawEntity, fact_hash_str) + `edge_kind.rs` (EdgeKind + all tests) — both
  under 500-line cap.
• `symbol_graph.rs` → `symbol_graph/` module directory: `graph.rs` (struct +
  save/load), `build.rs` (five build passes), `traverse.rs` (BFS), `tests.rs`
  + `persist_tests.rs`. `load_from_corpus` now delegates to `EdgeKind::from_tag`
  which implements Option H: `"custom:"`-prefixed tags round-trip, bare
  unknown tags drop + increment `unknown_edge_tags_dropped`.
• `save_to_corpus` and `edge_kind_breakdown` now use `EdgeKind::tag()` so
  Custom edges persist and report correctly.
• Fix Copy-removal call sites across `trusty-common/symgraph/graph.rs`,
  `trusty-analyze/types/graph.rs`, and `trusty-analyze/adapters/typescript.rs`.
• `GET /indexes/:id/graph/stats` already included `unknown_edge_tags_dropped`
  (prior PR); verified field is wired.
• Allowlist updated: `symbol_graph.rs` entry removed (file deleted).

Quality bar: cargo build ✓, cargo test -p trusty-common/trusty-search/trusty-analyze ✓,
clippy -D warnings ✓, fmt --check ✓, check_line_cap.sh ✓ (165 grandfathered, 0 violations).
